Rename `Main` to `Diary`
