+++ app/src/main/java/com/chaeniiz/weatherdiary/presentation/diary/DiaryView.kt
@@ -1,6 +1,6 @@
-package com.chaeniiz.weatherdiary.presentation.main
+package com.chaeniiz.weatherdiary.presentation.diary
 
-interface MainView {
+interface DiaryView {
     fun setWeatherTextView(weather: String)
     fun setContentTextView(content: String)
     fun startWriteActivity()
