+++ app/src/main/res/layout/activity_diary.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.main.MainActivity">
+    tools:context=".presentation.diary.DiaryActivity">
 
     <TextView
         android:id="@+id/weatherTextView"
